+++ packages/snap/hardhat.config.ts
@@ -27,6 +27,42 @@ if (MNEMONIC && INFURA_PROJECT_ID) {
       mnemonic: MNEMONIC,
     },
   };
+  networks.avalancheFuji = {
+    url: `https://avalanche-fuji.infura.io/v3/${INFURA_PROJECT_ID}`,
+    accounts: {
+      mnemonic: MNEMONIC,
+    },
+  };
+  networks.arbitrumSepolia = {
+    url: `https://arbitrum-sepolia.infura.io/v3/${INFURA_PROJECT_ID}`,
+    accounts: {
+      mnemonic: MNEMONIC,
+    },
+  };
+  networks.optimismSepolia = {
+    url: `https://optimism-sepolia.infura.io/v3/${INFURA_PROJECT_ID}`,
+    accounts: {
+      mnemonic: MNEMONIC,
+    },
+  };
+  networks.lineaGoerli = {
+    url: `https://linea-goerli.infura.io/v3/${INFURA_PROJECT_ID}`,
+    accounts: {
+      mnemonic: MNEMONIC,
+    },
+  };
+  networks.celoAlfajores = {
+    url: `https://celo-alfajores.infura.io/v3/${INFURA_PROJECT_ID}`,
+    accounts: {
+      mnemonic: MNEMONIC,
+    },
+  };
+  networks.polygonMumbai = {
+    url: `https://polygon-mumbai.infura.io/v3/${INFURA_PROJECT_ID}`,
+    accounts: {
+      mnemonic: MNEMONIC,
+    },
+  };
 }
 
 const config: HardhatUserConfig = {