+++ packages/site/CHANGELOG.md
@@ -7,6 +7,49 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0]
+
+### Changed
+
+- Make `@metamask/snap-account-abstraction-keyring-site` publishable ([#215](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/215))
+- Use new CentraNo1 font ([#184](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/184))
+
+## [0.5.0]
+
+### Changed
+
+- Use new `keyring-api` modules layout ([#166](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/166))
+
+## [0.4.2]
+
+### Changed
+
+- Bump `@metamask/keyring-api` from `^8.0.2` to `^8.1.3` ([#141](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/141))
+  - This version is now built slightly differently and is part of the [accounts monorepo](https://github.com/MetaMask/accounts).
+
+## [0.4.1]
+
+### Changed
+
+- Added default params value for walletInvokeSnap ([#126](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/126))
+
+## [0.4.0]
+
+### Changed
+
+- Better user feedback if selected chain ID is not correct ([#120](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/120))
+  - It now displays a red border in case of error
+
+### Fixed
+
+- Fixed issues with test IDs, context usage and state ([#120](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/120))
+
+## [0.3.0]
+
+### Added
+
+- Deployment of verifying paymaster from frontend ([#71](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/71))
+
 ## [0.2.2]
 
 ### Changed
@@ -29,7 +72,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.5.0...v1.0.0
+[0.5.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.4.2...v0.5.0
+[0.4.2]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.4.1...v0.4.2
+[0.4.1]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.4.0...v0.4.1
+[0.4.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.3.0...v0.4.0
+[0.3.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.1.0...v0.2.0