+++ packages/snap/CHANGELOG.md
@@ -7,6 +7,48 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0]
+
+### Changed
+
+- Bump `ethers@^5.8.0` ([#179](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/179))
+
+### Fixed
+
+- Return `null` if `handleKeyringRequest` returns `undefined`/`void` ([#180](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/180))
+
+## [0.5.0]
+
+### Changed
+
+- **BREAKING:** Provide `scopes` field to `KeyringAccount` during account creation ([#166](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/166))
+  - Bump `@metamask/keyring-api` from `^8.1.3` to `^13.0.0`.
+  - Compatible with `@metamask/eth-snap-keyring@^7.1.0`.
+
+## [0.4.2]
+
+### Changed
+
+- Bump `@metamask/keyring-api` from `^8.0.2` to `^8.1.3` ([#141](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/141))
+  - This version is now built slightly differently and is part of the [accounts monorepo](https://github.com/MetaMask/accounts).
+
+## [0.4.1]
+
+## [0.4.0]
+
+### Added
+
+- Add localhost chain ID to chain config ([#120](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/120))
+
+## [0.3.0]
+
+### Changed
+
+- Bump dependencies `@metamask/keyring-api` to `^6.0.0`, `@metamask/snaps-sdk` to `^4.0.1` and other minor versions of @etherumjs and @metamask dependencies([#86](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/86))
+- Fix chain config setup ([#62](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/62))
+- Fix undefined param for chain config ([#81](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/81))
+- Remove pending requests and associated methods ([#45](https://github.com/MetaMask/snap-account-abstraction-keyring/pull/45))
+
 ## [0.2.2]
 
 ### Changed
@@ -38,7 +80,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.5.0...v1.0.0
+[0.5.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.4.2...v0.5.0
+[0.4.2]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.4.1...v0.4.2
+[0.4.1]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.4.0...v0.4.1
+[0.4.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.3.0...v0.4.0
+[0.3.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/MetaMask/snap-account-abstraction-keyring/compare/v0.1.0...v0.2.0